Instead of cloning systems have extra maps
This new map pair helps with minimizing the amount of clones needed.
Cloning systems is much harder than simple IDs ofc.
The only bottleneck at this point however that might come up is that we
hash much more this way, which if we end up tampering with the hashing
function of a system might cause some trouble.