Binary heap approach
This is currently more like a DFS for finding any solution instead of
the best one. This is MUCH faster and scales way better for this problem
as well.
On the long run this should be extended such that the node value for the
minimum priority queue is some sort of reasonable heuristic. But in the
end this might not provide a shortest solution after all.